+++ src/main/scala/stellar/sdk/FederationServer.scala
@@ -12,9 +12,9 @@ import stellar.sdk.model.response.{FederationResponse, FederationResponseDeseria
 
 import scala.concurrent.{ExecutionContext, Future}
 
-class FederationServer(val base: Uri, path: Path)
+case class FederationServer(base: Uri, path: Path)
                       (implicit val system: ActorSystem = ActorSystem("stellar-sdk", ConfigFactory.load().getConfig("scala-stellar-sdk")))
-  extends WebClient with LazyLogging {
+  extends WebClient(base) with LazyLogging {
 
   implicit val formats = Serialization.formats(NoTypeHints) + FederationResponseDeserialiser
   import HalJsonSupport._